requirements.txt
affinegap==1.11
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
BTrees==4.9.0
categorical-distance==1.9
certifi==2020.12.5
cffi==1.14.5
cycler==0.10.0
datetime-distance==0.1.3
decorator==5.0.9
dedupe==2.0.8
dedupe-hcluster==0.3.8
dedupe-variable-datetime==0.1.5
defusedxml==0.7.1
DoubleMetaphone==0.1
entrypoints==0.3
fastcluster==1.2.3
future==0.18.2
fuzzywuzzy==0.18.0
haversine==2.3.1
highered==0.2.1
importlib-metadata==4.0.1
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
levenshtein==0.12.0
Levenshtein-search==1.4.5
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
persistent==4.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pyaml @ file:///home/linux1/recipes/ci/pyaml_1610972911625/work
pycparser==2.20
Pygments==2.9.0
pyhacrf-datamade==0.2.5
PyLBFGS==0.2.0.13
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
rlr==2.4.5
scikit-learn==0.24.1
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
simplecosine==1.2
six==1.16.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
Unidecode==1.2.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.1
zope.index==5.0.0
zope.interface==5.4.0
